Fix BreakpointEditDialog disappearing after changing the type
When the user selects a different breakpoint type (e.g. changing a
log point into a conditional breakpoint) we update the placeholder
of the input field.

This triggers a 'blur' event. The DebuggerPlugin closes the edit dialog
after receiving 'blur'. This is normally desired, as clicking outside
the dialog should close it, but in this case is closes the dialog
pre-emptivley.

This CL fixes this by focusing the editor before updating the
placeholder. This way no 'blur' event is sent and on top of that this
even removes a click for the user who would have to re-focus the
editor after changing the type.

R=dsv@chromium.org

Fixed: 1485782
Change-Id: I2b6bdd382ac9086f29053f986ca784e011a520d3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4896506
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>