Ensure Perf Panel search works with new engine
This CL makes a few changes to the panel's filtering logic to ensure
that the search functionality works when we are using the new engine.

The issue was that we converted new engine events to legacy events for
the sake of filtering, but the filtering code should all accept both new
and old trace events, so we don't need to do that.

Then we need to make sure that if we are looking at a profile node, that
we look at the call frame as the text to use for filtering. I then went
a bit further if we have sampling data, because we can then use the name
from the ProfileNode that represents the entry. This is good because
this is where the name gets updated should we have sourcemaps and have
resolved the function name, so that is the one we want to use if we can.

There is more work to do here across all the filtering logic - it is a
bit hard to follow and could probably be structured differently - but
that is not going to be impactful work now and is likely something we
can tidy once the old engine has been removed.

Fixed: 1484504
Change-Id: If6baf135acd813ff82ca062b07c1dfd63dbe964d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4900636
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>