[coverage] Fix line-by-line coverage for pretty-printed code
This CL changes CoverageDecorationManager#usageByLine to support
UISourceCodes where the content was formatted. We do this by
providing a line mapping to usageByLine. The mapping is an array where
each element represents a line and it describes where in the
UISourceCode#content that line can be found.

The rest of the code stays as is. This even works if the UISourceCode
content comes from a source map, as the source mapping step happens
after we "reverse" the formatting.

R=jarin@chromium.org

Fixed: 1464974
Change-Id: I82b54c3ca52deee4ce3894b601696b9cce8ece53
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4905366
Reviewed-by: Jaroslav Sevcik <jarin@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>