Do not filter LCP based on the calculated active window of its process
We are incorrectly calculating the active windows for processes (see
https://crbug.com/1487950), so filtering LCP events based on a window
that's not correct is probably not something we want to do. Besides, we
are already filtering LCP events to only the outermost frame, so this
filtering is likely something we didn't need in the first place.

Note: The original plan was to remove the window calculation logic
completely, but there is a dependency by the network handler which I
had missed on it so we cannot remove it yet.

Fixed: 1487136
Change-Id: Ibb6f0e0fad4e49acee5a668b7625b80b2a6ee955
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4901985
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>