Don't drop AuctionWorklet processes in the RendererHandler
In order to render the processes, we need to keep the events around.
Previously these processes were dropped due to a lack of URL, but this
CL updates the RendererHandler to keep them and use the
AuctionWorkletsHandler to set the URL, thus maintaining them.

Note that this CL does NOT use the new engine to render these tracks, I
want to do that in a follow-up CL. Therefore in
CompatibilityTracksAppender we currently skip a process if we detect it
is for an AuctionWorklet.

Bug: 1478710
Change-Id: Ieba0cfbbcccc59a0b9fa830a7f5074248122ad88
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4907927
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>