Fixed Narrator not announcing checkbox check or unchecked under 'Sources > Breakpoints'
Narrator does not announce that a checkbox is checked or unchecked when clicking a checkbox under "Sources > Breakpoints".  This is because the control type is missing on the div even though it is clickable.  This change adds role='checkbox' and 'aria-checked' to the div.

Before:
https://imgur.com/a/VDOpe7J

After:
https://imgur.com/a/7oXgIkR

Bug: 1488645
Change-Id: I5de07e5968b249674f226336ceb22c697447fe15
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4908618
Commit-Queue: Jimmy Seto <jimmyseto@microsoft.com>
Reviewed-by: Danil Somsikov <dsv@chromium.org>