Allow TraceLoader to take TraceEngine config
This CL enables the caller to pass TraceEngine configuration options
when using the TraceLoader to parse a trace in a test. Previously we
only cached trace loader calls on the trace file name; this CL updates
that to also test based on the configuration.

To avoid JSON.stringifying the entire config, a function has been added
that can generate a cache key for a given config. This also avoids
issues like ordering of keys causing a cache miss.

This is useful because I want to add some tests where I change the
settings and assert different behaviour in a handler; I cannot do that
without this CL as the resulting data will be cached regardless of the
config settings.

Bug: none
Change-Id: I40589a8f9d57b0974c9559a1ce9f9ee445e14340
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4907937
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>