Enable Tab target for Puppeteer
This CL changes the Puppeteer instance used by the e2e and interaction tests to work in the Tab target mode, that is the mode that supports MPArch and features like pre-rendering. In the next releases of Puppeteer it will be enabled by default and this configuration can be removed.

Bug: chromium:1484476
Change-Id: I771234935a1d95de143a6c4559aa297d7dd1b4ae
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4853353
Auto-Submit: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>