Simplify new engine color definitions and use M3
I noticed we are showing wrong colors for profile calls in the timeline
tree view. When looking for a quick fix for this, I realized that we
could benefit by unifying the styles definitions as much as we easily
can. This is what this CL does.

Specifically, this takes TimelineRecordStyle and TimelineCategory
and moves them into EventUICategory.ts (the file we use for events of
the new engine). Inside this file, I removed the classes that emulated
what these two do already and adapted the implementation to make it more
straight forward to use.

Note this doesn't deduplicate the mappings from event name to the
corresponding styles, since these rely on the types of both engines and
require more work to do so. (We have decided to remove the old engine's
mapping once it's no longer needed).

I took the opportunity to use the new M3 colors in the new engines
mapping.

Bug: 1428024
Change-Id: I781e7752fa29a41385dd1fe3dbdf9ccb57200e96
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4909540
Commit-Queue: Andres Olivares <andoli@chromium.org>
Auto-Submit: Andres Olivares <andoli@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>