Fix ProfileCall nodes being excluded
The filters we apply when generating the tables were incorrectly
removing any ProfileCall events due to them not having a matching record
type. To fix this, we map their RecordType to the RecordType.JSFrame
type, as they are broadly equivalent.

Long term we need to figure out what to do with this RecordType, and how
to reconcile it with the new engine, but this is a fix that enables the
bottom up chart to be correctly populated.

Bug: 1488757
Change-Id: Ic394371e683251c6c77699210bcd70a62be16e3b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4909533
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>