Fix console drawer extra metrics when initialized
Currently when the drawer is first initialized via choosing a specific
drawer to open, the "initialization" will first default the drawer to
the Console tab, sending an extra panelShown metrics before showing the
actual drawer tab that user selected. This CL fixes this by preventing
the auto-select when a specific drawer is targeted, and only turns on
the auto-select when this initialization action comes from
"Open console drawer", which does not have a dedicated drawer to open.
(It opens the last state of the drawer, and if unopened previously,
default to the Console).

Fixed: 1134521
Change-Id: I7f370ccd12ce66f9b0c794e2c1d6c6ea30c535a9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4952374
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Changhao Han <changhaohan@chromium.org>
Auto-Submit: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>