Add Thread helper for new trace engine
This CL adds a Threads helper that can be used to gather threads
regardless of if the current trace's source is tracing or CDP profiling.

Annoyingly, because it needs to refer to data in the handlers, it cannot
live in the Helpers module (it would create a circular dependency) but I
think it makes sense existing as a module in the Handlers.

Note that this function is not used anywhere, I will follow-up with CLs
that introduce it and also tidy up the duplication (ThreadType enum is
now duplicated with the CompatTrackAppender).

Bug: 1492115
Change-Id: Ibc5ebc7e0b54538d28c37cc74ba7e7535e02a6b4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4965362
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>