[Overlay] Restore persistent overlays between DevTools sessions
The idea is similar to how we keep DOM breakpoints:
* Saving a local setting on state change and keying the state with document URL and the node's path.

I needed to move callbacks to the persistent highlighter because we need to dispatch events while restoring the state.

Updated ElementsTreeElement to check for overlay state while adorning elements so that if they are being created after overlay model is initialized and restored; the adorners will be toggled on in the Elements panel.

Drive by:
* Removed `#flexEnabled` parameter as it is not used in anywhere.
* Updated the mocks in the tests to use sinon.

Fixed: 1221486
Change-Id: I111c49f31a5dfcaeb95c2e4c997953e919d9f07c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4952377
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>