Support ignoreList field in source maps
During source map standardization discussions, there has been a
unanimous desire to rename x_google_ignoreList to ignoreList and
make it part of the spec. It is now in the spec, so this change
implements support for it, with fallback to x_google_ignoreList
if ignoreList is not present.

Bug: b:288553764
Change-Id: I1a2dc7f6b60bb7af0c825b01ba3d3afb729f220e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4966546
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Eric Leese <leese@chromium.org>