Fix Timeline MiniMap not redrawing CPU activity
This CL fixes a bug where CPU Activity would not be redrawn correctly
respecting the left and right custom window it was being asked to draw.

The fix is to allow our tree walker helper to filter events by a range,
and that way we ensure only those events are used to generate the
minimap.

As a drive-by fix, this CL also fixes the selection colour used when
dragging on the minimap, which was mistakenly set to a opaque colour and
not one that is transparent.

Bug: 1464206
Change-Id: Iec3589621c2c9f1c3028b943611a9c66cbbc6631
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4976447
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>