Use resolved source map names in bottom-up grid
We were using them in the timeline, but not in the bottom-up grid.

I also realised that rather than find the right node for an entry each
time, we could instead maintain a map of ProfileNodeId=>string|null
which represents the resolved function name.

This CL creates and uses that map for the bottom-up grid, but there is a
good opportunity here to tidy up and create a
`Helpers.functionNameForProfileNodeId()` helper which we can then use
everywhere, and it will prioritize the resolved name before falling
back. I will work on implementing this in a future CL, but this CL is
purposefully kept small to fix the bug where we do not use resolved
names in the bottom-up grid.

Fixed: 1497371
Change-Id: I80f63f8e06d89e86d58bfff845b586f4d0c38858
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4987624
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>