Balance unmatched left parentheses when submitting style text edits
Currently in the Styles pane, when a `(` is entered, for example, from
`calc(` or `var(`, the `)` is not automatically followed, which means
that if the autocomplete suggestion is accepted and the cursor moves
to the next line to create the next declaration, the `(` is often
unmatched, creating an invalid declaration. This CL fixes this behavior
by fixing unmatched left parentheses when submitting the text to the
backend.

Fixed: 1294251
Change-Id: I286b54f8fd63c0566cba495a294dc92bd7690775
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4987627
Auto-Submit: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Changhao Han <changhaohan@chromium.org>