Fix missing console.restore for interactions
This CL fixes a case where the canvas would suddenly have its font drawn
in much larger sizes.

The code for interactions, if the text did not fit inside the container,
would early exit to avoid drawing the text. However, when it did this,
it stopped a `context.restore()` call from running, which left the
canvas state in a bad situation.

The fix here is to ensure we always restore the context. As a drive-by
improvement I also removed an extra set of restore() and save() calls,
so the entire `drawInteractionEventWithWhiskers` function is now only
creating and restoring one context.

Bug: none
Change-Id: I0e63b197c7023a59cac1ad7100bd02324381d0d3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4993226
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>