RPP: shift breadcrumbs button depending on screenshots
In crrev.com/c/4991521 I moved the breadcrumbs up because they collided
with the minimap border, but I realised that if screenshots are enabled
it looks better to have the button a few pixels lower.

Therefore this CL adds a class when screenshots are enabled, and shifts
the button based on that. This means it does not overlap when
screenshots are off, but also is positioned better when they are on.

Bug: 1497893
Change-Id: I774b2282447314d4013b50540ed16a4be8d1ffcf
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4994348
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>