[RPP] Use computed color value in category nodes in the tree view
We were still using the css variable names on these particular nodes and
as a result the color on their chips were not being rendered. See
screenshot.

Drive-by fix, dispatch a NamesResolved event every time names are resolved, including when the resolver is installed (so that we automaticaly show resolved names if they are resolved after the first flamechart render).

https://screenshot.googleplex.com/4k6XjiUKSuKpbn4

Bug: none
Change-Id: Ia4d075218796dd24361621d09bc89ed00c8a617b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4995191
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>