[console] Make Ctrl+` behave like in VS code.
Prior to this CL, Ctrl+` focused the Console panel or drawer when
something else was focused; but when the Console drawer has focus
it didn't behave like in VS code and did not close the drawer.

With this CL we now properly close the drawer if the Console in the
drawer was focused when the developer pressed Ctrl+`.

Fixed: chromium:1379256
Change-Id: If9688319378a00da4d5778910b8b3c9ce24deecc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5001007
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>
Commit-Queue: Yang Guo <yangguo@chromium.org>