[RPP] Extract stack trace from new engine trace events
Trace events that contain a stack trace (f.e. FunctionCall events) can
include a link to the corresponding call frame in their detailed view.
At the moment we were only supporting this for profile calls, so this CL
adds support for trace events with a stack trace.

Note we are still missing support for events for which a stack trace is
artificially built. This is only happening in the old engine (as part of
the timelineData logic) and as such should be done when that logic is
ported to the new engine.

Bug: none
Change-Id: Iba8373f9e3abfff1a1c4359fbb2085e7c8f1d030
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4995193
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>