Fix some of the computed style properties are not rendered.
It seems like we're not rendering when the default values doesn't change.
So, for properties that have the same values as default, the element
is not rendered into the created shadow root.

This CL calls initial render when the custom element is connected
to the document.

Fixed: 1499882
Change-Id: I8aab31918741220845bf43b2ac5a621e4e64d15e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5003862
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>