Fix call stack RangeError in WorkspaceImpl.ts
This CL replaces spread notation with a for-of loop in WorkspaceImpl.ts.
We ran out of JS stack space when debugging DevTools-on-DevTools with
a debug build due to the number of UISourceCodes in a single project.

R=bmeurer@chromium.org

Bug: None
Change-Id: I4ffcb6d0bca1804f5a8fb0595aca1715fbd9465c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5012877
Auto-Submit: Simon Zünd <szuend@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>