RPP: improvements to the minimap redrawing
When the breadcrumbs experiment is enabled the minimap has to be redrawn
as we zoom in and out. The existing minimap implementation is not really
designed for this and as such redraws can be very slow.

There is more to do here before we can ship breadcrumbs into production
but this CL makes a few improvements:

1. Avoid using the event timing helpers which cause a lot of objects to
   be created and then GC'd.
2. Add a minDuration to the walkTree code - tiny events will make no
   visual impact on the minimap so we can skip them.
3. Cache the creation of categories for entries - this is relatively
   expensive to create and once we have done it for an entry once, we
   can cache it and know that it will not change.

Bug: none
Change-Id: Ib02a76e0d46c6af4a2e8eded98fb474348d4abf1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5001471
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>