[RPP] Remove the |flameChartData| from *TrackAppenders
Instead, we directly get it from the |compatibilityBuilder|.

The reason is for flame chart transformation, we will create a new flameChartData object, so the reference in the appenders will point to the old object.
To resolve this, we should remove unnecessary reference and (long-term (crbug.com/1501055)) try to change in the same object.

Fixed: 1500976
Change-Id: I16d35240362c7c0034f8a316c9e575f252fa2a84
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5018336
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>