[Flame Chart] Update the group tree instead of re-rendering it if the tracks are not changed
The "not changed" here means the order of the tracks, but the starting level and other properties can be changed.

After this change, if a use reorder the tracks, the new order will be kept until a new profile is loaded. This will be very useful for the flame chart transformation, like merge an entry in the main track, it will keep the scrollTop/Left of the origin view.

Bug: 1469886, 1495583, 1500623
Change-Id: I6cc2c004b2c923b03eadbbf4263c38b4966031ec
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5018339
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>