(frontend 2) Preloading DevTools: Show header mismatch on prerender activation
In the previous CL[1], it changes the protocol on the DevTools side so
that DevTools can receive the information of mismatched headers.

This CL enables DevTools to display mismatched headers in a table on the
Speculative loads panel when the header mismatch occurs.

[1] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4930959

Demo:
https://drive.google.com/file/d/1lyEvlnJ_b27ux_Usaa33tOKrFEqMEsDL/view?usp=sharing

Design Document:
https://docs.google.com/document/d/1lIwZwBMgCQXH0hRnc2Guj-zOBI2rHtkI2yinhYjsFvo/edit?usp=sharing

Bug: 1456673
Change-Id: If58fde28621b844e378395cda721db12cd9ea109
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4933567
Reviewed-by: Ken Okada <kenoss@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Lingqi Chi <lingqi@chromium.org>
Commit-Queue: Miina Koyama <miinak@chromium.org>