[cleanup] Replace ActionRegistry#action() with type-safe alternative.
Until now we have been using `ActionRegistry#action()` with two
different contracts: On the one hand as a way to get a pre-registered
action, where we just cast the result to `Action`, and on the other
hand to test if a given action is registered. This has been a source
of surprises, especially since in the first case there was no meaningful
error that you are holding it wrong.

This CL replaces `action()` with a pair of methods `hasAction()` and
`getAction()`, and the latter throws a meaningful error if the action
isn't found.

Bug: chromium:1467464
Change-Id: Id507a3280ea1f780e77c4216a567fd297a04a77b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5033626
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>