[RPP] add InitiatorsHandler
This CL adds the first few initiators that we detect in the old engine
and represents them in the new engine.

The logic and tests here come directly from inspecting and following the
old engine; I have tried to add comments through where I can in order to
make it easier to follow.

This is not complete; the follow-up steps are to implement the rest of
the initiators (there are a bunch of async ones - I think we can reuse
some of the old engine's implementation). Once that is done we can then
update the UI to look up these initiators.

Bug: 1434596
Change-Id: I60926fa7c36d88a4012f3d798b483074c999ec08
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5040795
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>