[RPP]: create new frames model
This CL updates the new Trace Engine to create an instance of the Frames
model when parsing a file.

It's slightly unfortunate that we cannot easily use a handler here and
this breaks the pattern, but the TimelineFramesModel is a large bit of
functionality that right now we don't want to have to completely
rewrite.

Bug: 1499182
Change-Id: If4fe9be3d88e82b90bcf17eefd974909115655f0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5049781
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>