RPP: move tooltip warnings up to CompatibilityTracksAppender
Previously this logic lived only in the ThreadsAppender because it was
the only track that had warnings associated with it. But this is not
true now because of interactions, and long term any track could have
warnings.

Therefore, rather than add warnings per-appender, we can instead
generate the code at the CompatibilityTracksAppender level, rather than
duplicate calls to it in each appender.

Bug: 1496750
Change-Id: I366055fbe1a10ea4dda65cc8cb6ab068b6986bc1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5050081
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>