RPP: Remove old engine warnings
This CL removes the old warning system; all the frontend is being
rendered and powered via the new engine which has the warnings we want
to keep in the WarningsHandler.

There is a bit more code to deal with and delete following this CL -
primarily the warnings key of EventOnTimelineData, but I will do that in
a follow-up.

Bug: 1496750
Change-Id: I6a4326ecf47aac5477b02086f0bd1e8445e4a4e5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5050082
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>