Migrate TimelineFrameModel to a handler
When beginning to work on using the migrated TimelineFrameModel I
realised that it not being a handler was going to cause issues - I would
have to begin to pass around the model instance rather than just the
parsed handler data.

I then realised that with a bit of work I could wrap the class up in a
handler, and return the set of frames from the handler.

This CL does just that. In the future we can slowly migrate this away
from the big class and into logic that is more conventional with regards
to the other handlers, but that does not seem worthwhile right now.

Bug: 1499182
Change-Id: Id4bb83f6242492bc40ef1fc7fdc422258df9a58e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5050209
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>