RPP: Begin removing ThreadTracksSource
This CL begins to remove more of the legacy engine starting with
ThreadTracksSource. This doesn't get rid of all of it - I didn't want
the CL to get too large, so this is the first part. I will remove more
in the next one.

Now that we know this source is effectively always the new engine, it
lets us begin to strip down the PerformanceModel and TimelineModel. For
example, this CL removes the old source map resolution code in
PerformanceModel as it is not used anymore.

Bug: 1499182
Change-Id: I0d85eca893f1dd20bcc1cea7aaf7bd06022c4d15
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5054329
Reviewed-by: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>