Remove EventOnTimelineData.initiators
This CL removes the legacy tracking of initiators.

It is important to note that this CL removes some initiator logic
relating to network events; this logic doesn't exist directly in the new
engine but does in some part because we use all the related network
events to create the synthetic network event containing all the
information. Therefore we do not need to track initiaors on each part of
the network request.

Bug: 1434596
Change-Id: Ifcc12eb7a5b4a0de4377a2edb1a7904a30a3de1a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5059269
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>