[RPP] Create new breadcrumbs when new trace is loaded or chosen
Setting this.#currentBreadcrumbTimeWindow to undefined in TimelineFlameChartView. Previously, loading second trace was failing because TimelineFlameChartView assumed breadcrumb exists when it belonged to a previous trace.
Also, moved adding initial breadcrumb to the method that is called every time a trace is loaded of switched to. This way we are creating new breadcrumbs object each time.

Bug: 1505028
Change-Id: I7462276e1f7308a924442ceacc6300afa6d21907
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5073672
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>