Use well-formatted related code
If related code is on one line, we now pretty-print it first. Also
we use the indenting of the code to pick a range that is likely to
be well formatted, in that it will probably not contain the start
or end of a block unless it contains the entire block.

Bug: b:299423255
Change-Id: I347deb42135dfb0148a40a495560e3c41d7bad73
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5077377
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Eric Leese <leese@chromium.org>