[Overrides] Fix logic for has-overrides indicator
A request can be intercepted, but not be overridden (neither content
nor headers). One scenario where this happens, is when there are
overrides defined for a resource with a very long URL, which can lead
to other related URLs being intercepted but not overridden.

This CL fixes the logic for the purple dot indicator in the network
panel, to take this case into account correctly.

Fixed: 1504956
Change-Id: I30d7b190f8d6316827b123c07002ed59963978ac
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5079479
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Jecelyn Yeen <jec@chromium.org>
Commit-Queue: Jecelyn Yeen <jec@chromium.org>