RPP: gather all trace events in an array
We rely on TracingModel.allRawEvents() to gather all the events, but the
plan is to remove this model. Therefore the loader needs to store all
the events in an array so they can be passed back into the panel and
processed by the trace engine.

Bug: 1499182
Change-Id: I60af2178404c5ee16726cb9b5c39a3bce9f18f77
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5094042
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>