[RPP] Remove individual Context Menu undo actions
We changed the scope of this feature to only be able to unhide all children of an entry so this CL deleted code for removing individual actions.

There will be at least 1 other Undo action added in the future, therefore FilterUndoAction enum was not deleted.

Bug: 1469887
Change-Id: Id00b1ce3e9c824b6b3b25bb85efecf1d1668dfa3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5104934
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>