Fix media tab test flakes
The test was flaking because the async methods were only waiting for the
kPlay event, and the promise for this was being resolved before the
kLoad event was triggered. Now `playMediaFile` awaits for ended,
ensuring that all the events which change the devtools state have been
fired.

Bug: 1368558
Change-Id: I2f6982291890c8cce1dea72f39cb4ba551cd623b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5114029
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Ted (Chromium) Meyer <tmathmeyer@chromium.org>