Trace: deal with wrong URL in TracingStartedInBrowser
Sometimes in the trace file the TracingStartedInBrowser event will show
that the main frame's URL is either about:blank or the previous URL
rather than the one that was actually recorded. This doesn't happen all
the time (not sure on the exact trigger) but can happen. In this case we
can look at the first main frame navigation and use its URL (if we got
a navigation in the trace at all) to be a little bit more accurate.

Bug: none
Change-Id: I58b7b7b95235a92f470f5f1f46db690de72d78d5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5116653
Reviewed-by: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>