[sources] Properly align revealing of sources with VS code.
The behavior for 'Automatically reveal files in sidebar' is now:
When enabled, the navigator view in the 'Sources' panel will always
switch to the relevant tab for the source file and will reveal the tree
item. But it will never automatically open the navigator view if that
sidebar was closed. It will also memorize the last open tab within the
sidebar, independent of whether that tab was selected by explicitly
clicking on a tab header or via revealing.

When explicitly revealing a file in the navigator view from the context
menu, the behavior is the same as for automatic reveal, but in addition
the sidebar will also be showing if that's not already the case, and the
focus will be moved to the navigator view (aligned with VS code behavior
in that scenario).

We also add an 'Reveal active file in navigator sidebar' action, similar
to what VS code offers, and for consistency we rename the 'Reveal in
sidebar' context menu item to 'Reveal in navigator sidebar', which is
consistent with the UI text we use for toggling the sidebar.

We use 'sources.reveal-in-navigator-sidebar' consistently as jslog
context for this.

Bug: chromium:1467464, b:308382451
Change-Id: I10f9f664bb92338b3df51af30c83830905e3da8f
Fixed: chromium:1509387
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5119545
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>