RoutingDevTools: Take router's ID from ServiceWorkerVersion in frontend
Currently, DevTools frontend calculates the ID of routers by itself.
But router IDs are originally generated in the Chrome backend, and it
is not preferable to have multiple codes doing the same thing. This CL
is a frontend change to fix this issue.

Backend: https://crrev.com/c/5119403
Bug: 1511169
Change-Id: I01c870d4ba14c675bc5e2a3334682dc44b448529
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5125354
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Commit-Queue: Yoshisato Yanagisawa <yyanagisawa@chromium.org>