RPP: fix generic traces in the new engine
I discovered that generic traces were broken when we used the new
engine.

The root cause of this was the RendererHandler, which used
processesPerFrame to generate a list of threads - but in generic traces
we might not have that information, so now it loops over all the
processes it finds. I don't think this causes breakages - it just means
for generic traces we are more accurate. We also avoid culling by URL
for generic traces now also.

Aside from that this CL contains some minor fixes to handle generic
traces, and also extends the MetaHandler to expose a map of ProcessID to
process_name events so we can name generic processes nicely.

Bug: 1499182
Change-Id: Ia9520c3aedbb0eae0bf2612dcdb2de0afc724977
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5140430
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>