[ssci] Update Lighthouse metadata
This CL updates the URLs for Lighthouse and Lighthouse i18n.js, so that
it is an absolute URL and successfully links in chrome://credits
instead of being a relative link.

Also specified the Version for Lighthouse, as at least one of Version,
Revision or Date should be specified for third party metadata.

Bug: b:290296624
Change-Id: Ib31c5e9ad5f555a0100ebf3591b1c094aaf80294
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5162108
Reviewed-by: Connor Clark <cjamcl@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Paul Irish <paulirish@chromium.org>