[DevToolsIcons] Use the `--image-file-<name>` :root variables.
We already define `--image-file-<name>` variables for all the images
supported, with the correct URLs, so we can as well just use those
directly instead of trying to construct a URL manually in another
place. This also has the advantage that it's easier to see in
DevTools-on-DevTools when an image file is missing.

Bug: chromium:1514825, chromium:1421730
Change-Id: I870b35f199cda472428d85a4611052db905e6857
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5166407
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>