[application] Fix text coming out of popup when deleting a database.
When we try to delete a DB in indexedDB inside storage section in
application in inspect element, the confirm dialog box and its text
was popping out of screen when we drag the inspect element to right
increasing the page view.

Updated some CSS in `confirmDialog.css` file to fix the issue.

Fixed: chromium:1514335
Change-Id: Id90a911d845b993a2e8fc17aa13a32fd270c501e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5163750
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>