[cleanup] Improve testability of the `Revealer` logic.
This adds a new `RevealerRegistry` (singleton), which manages the state
and the lookup of the registered revealers, and allows for unit testing
with sinon. This removes the current pitfall of the `setRevealForTest`
and the global `registeredRevealers` that easily leaked between unit
tests.

Bug: none
Change-Id: I9ea14dc1f4b2e8b9e8a38a1711b4be7d3c724f6d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5180310
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>