[elements] Extract property names when ast matching
This fixes a bug where we'd not understand the property name when
there's a comment inside. The property name is parsed separately from
the value so that we're able to still interpret one or the other if
one of the two is malformed.

Bug: chromium:1504820
Change-Id: Id0f0d73b81190a2984d497d8e09f1db20c46a24d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5184395
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>
Auto-Submit: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>