[elements] Use authored text for editing styles
Previously the initial text used for editing styles was the textContent
of the respective name or value html elements. Since we started
normalizing whitespace this text now deviates from the authored text.
This CL fixes that and also refactors the editing implementation a little.

Bug: chromium:1504820
Change-Id: Idaf9ced5d133083421c77c210d0f5ad87b5ba419
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5184457
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>