[e2e] De-flake overrides tests
Following up on [1], this CL not only asserts that the "Override
content" context menu option is enabled, it also closes and re-opens
the context menu if it isn't.
The CL extracts this into a helper function and uses it in a few additional tests.

[1] https://crrev.com/c/5185516

Bug: chromium:1502463, chromium:1509276
Change-Id: I090e0a7e5c4e4c8074e3f5ab61ea6cb0a6933b79
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5198095
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>