Only add reflow warnings to long AND forced reflows
We were adding warnings to RecalculateStyles and Layout events if any of
them exceeded a threshold. However this is not what we want.

Instead we want to mark long forced reflows, that is, those
RecalculateStyles and Layout steps that are invoked by JS (sync) and
that when combined exceed a threshold.

Fixed: 1519941
Also-by: paulirish@chromium.org
Change-Id: I3ceaa2ae36e62c3f921d091c387c052fac68ca6d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5224682
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>