Improve computation of distances for WeakMap values
The value half of each (key, value) pair in a WeakMap is shown in heap
snapshots with edges pointing from both the WeakMap and the
corresponding key object, to indicate that the WeakMap retains the value
as long as the key is alive. This is useful, but can lead to a sort of
bizarre situation in the Retainers view if a value object retains its
own key, and is itself alive due to a long retainer chain. In that case,
the distance of the key (at the distance of the WeakMap plus 3) could be
less than the distance of the actual retainer path for the value, which
makes it look like the value retains itself via the key.

To fix the problem, we can increase the distance of the value object so
that it is at least the maximum of (key distance + 1, WeakMap table
distance + 1). With this change, the repro case in chromium:1290800
would list the WeakMap first, then the longPath retainer chain, and
finally the key. This is not quite ideal yet, but is an improvement from
the current behavior, which lists the WeakMap first, the key second, and
the longPath retainer chain third.

Bug: 1290800
Change-Id: I0f48c0ef13a534963714bfcc4544891f94bae49e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5246505
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Seth Brenith <seth.brenith@microsoft.com>