Fix RequestResponseView when displaying non-ASCII base64 encoded text
This CL fixes the RequestResponseView for the cases where text content
arrives base64 encoded in the frontend and the text is not plain
ASCII.

The new ContentData can handle this correctly but the DeferredContent
interface used everywhere looses the charset information. Luckily
NetworkRequest already uses the new ContentData, so all we need to
do is change the conversion function when going from a ContentData
to a DeferredContent.

Instead of preserving the existing behavior and pass the base64
version of the content, we now always decode and construct the
DeferredContent with the text version of the content.

The problem is that the other places use the DeferredContent to
construct data URIs. With base64 this works, but if we want to
construct valid data URIs from arbitrary text content we have
to apply escaping.

R=dsv@chromium.org

Bug: 1509336
Change-Id: I7e28b09ec3fd3cacc9c82c06ebd0d880ab3cd19b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5246387
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>