Add 'Reset children' to 'Possible actions'
We uses to check if it is possible to reset children by checking if an entry has arrow decoration. Arrow decoration is only added to entries that actions were invoked on. This way 'reset children' would not appear in the context menu of entries that have hidden children, but the action was not applied directly on them.
Instead, each time when opening context menu, let's check if an entry has any hidden children and show the option to reset them.

Bug: 1469887
Change-Id: Ic3f86cc31597be42ccc6176b61c79f30fc01d06a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5266339
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>