[ve] Instrument popovers
This CL adds visual element logging support for the various popovers
in DevTools. Since they are added pretty close to the document root
in the DOM we need a parent provider. Luckily the popover helper
requires a container in which to operate (e.g .for events), so we
use that.

R=dsv@chromium.org

Bug: chromium:40276793
Change-Id: Ia50b4d796da3339936307c7d13a7371da3e286cc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5272841
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>