Gray out cycles to the start object in the Retainers pane
In the Retainers pane, any object which is already opened at a higher
level of the tree should be shown as gray and not provide the option to
expand it further, to prevent the developer from accidentally traversing
cycles. This functionality doesn't work, however, when the cycle is to
the very first object which the user had originally selected. This CL
sets the snapshot node ID on the root node of the Retainers pane so that
findAncestorWithSameSnapshotNodeId can correctly determine when there is
a cycle back to that starting node.

Bug: 40805497
Change-Id: Ic9501944fa6d48855f055516839a6891d2a2c0e6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5273876
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Seth Brenith <seth.brenith@microsoft.com>