[RPP] Add "Initiator for" field to the events that are initiators
Previously, we had a link to the initiator, but not the other way around.
This CL adds a link to all of the events initiated by some event.
https://screenshot.googleplex.com/8LCEKcVF5BqGG5y

Also move stack trace information to be appended after the "initiator for" and "initiated by" fields.

Bug: 325024819
Change-Id: I377c01c42ebe153783ca6355deba00569f783d0e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5286574
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>