[ve] Report disappearance as resize to 0.
We generally don't track when VE disappear which is a conscious design
decision making the implementation simpler and more efficient.

However this poses some difficulties for the generic logs analysis I
have in mind: at every moment we can say that user finds themselves in a unique state
defined by a set of VE seen so far. This is good for understanding the
user journey in details, but too specific for a generic user analysis.
Therefore we should be agregating these states based on top K largest VEs.
This works, except we can't tell the difference between the panel user
is seeing right now and the panel they saw earlier. It is still
important that they saw something earlier for a detailed analysis but
not for generic one. To solve this, we can extend the notion of a small
VEs to the invisible VE and consider them to have the size 0.

This comes in handy in our implementation of panels where we don't
remove previosly seen panel from the DOM. User might have interacted
with one panel, made some changes, switched to another panel and then go
back. If we were to log disappearance and a new impression, it would be
hard to associate the VE interactions they had before the switch with
the ones they did after. Logging resize to zero gives us enough
information to understand that VE is not visible without loosing the
history.

Logging disappearance of only those element with 'track: resize' is
somewhat arbitrary but in practive is exactly what we need: we don't
care neither about resize or disappearance of say tree elements but we
do care about these events for a panel or pane.

Bug: chromium:1467464
Change-Id: I52c2658f33bd9496872926736fbcd5a82bcb0963
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5309834
Auto-Submit: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Danil Somsikov <dsv@chromium.org>