[RPP] Add render blocking requests insight
This contains the basic render blocking request information that will
eventually be surfaced to the user in RPP.

This does not include any estimated savings yet because we need to port
Lighthouses page load simulation models (Lantern). We plan to do this
in a follow-up pass.

Bug: 325064016
Change-Id: Iea01bb1ded61e78a39a76c9c7805007a830a581b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5250877
Reviewed-by: Andres Olivares <andoli@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Adam Raine <asraine@chromium.org>