[RPP] Draw the initiator arrow to/from the closest modified ancestor when the entry is hidden
We have an ability to hide entries through the flame chart context menu. This currently causes a bug when we are trying to draw initiators, but the initiated entry or the initiator are hidden.
This CL solves this by drawing the initiator arrow to the closest visible modified ancestor.
For that, we need to provide lists of hidden and modified entries (the ones that have an arrow on them because the user applied actions directly on those entries or it is a parent of a hidden entries) to Initiators.js that builds the initiator pairs.

Remove cache for event indexes because when we hide entries, the indexes of the same events change and the cache contains wrong indexes.

Bug: 326916488
Change-Id: I9f4ae9a6695c8f41c8c9a1686601ce06273b68c9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5323980
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>