[sdk] Add 'peekVLQ' to the source map iterator
To decode generated ranges of the scopes proposal we need to look-ahead
one VLQ number in the input stream. To this end we add a `peekVLQ`
method to the iterator.

This CL also renames the 'StringCharIterator' to 'TokenIterator' and
'decodeVLQ' to 'nextVLQ'. This conveys the intention better.

R=pfaffe@chromium.org

Bug: 40277685
Change-Id: Iba443b2fb1c9bfd34674d2e95edf6c8c7be8e15c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5331781
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>