[ve] Allow redirecting keyboard events for logging
This is needed because the legacy data grid does not support real focus
inside it. See crbug.com/1394518 for details.

To do this, shift some of the parsing of track directive to the
LoggingConfig and use stroger types.

Bug: chromium:1467464
Change-Id: I30d9a2deaa85e720cc5b2c04f5c8a40b6d45a3c0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5361993
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>