Improve coverage tool stats calculation for filtered url
The stats calculation for filtered url in coverage tool was not accurate as the tool uses the visible parent to do the calculation while the parent url may not match the specified filter.

This CL ensures that the tool is using the filter matching children data to do the calculation in the above scenario.

Example: https://imgur.com/a/rfsNWoe
 
Bug: 329253687
Change-Id: Ibaf88c2a6b0fa7a6de61efca68fb99d6c9aae833
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5368380
Commit-Queue: Jayson Chen <jaysonchen@microsoft.com>
Reviewed-by: Eric Leese <leese@chromium.org>