[RPP] Add AnnotationsManager and save Hidden Renderer Trace Events hashes into annotations metadata
Add a service that builds annotations to be saved into the metadata.
At the moment, only the Renderer Trace Events are saved.

The Annotations interface stricture is not complete and not yet being filled. I will add tests for generating correct annotations when it's generated with all of the data.

Bug: 313757110
Change-Id: I9cbf26549ed8b870daa9170609850526667d4a1d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5352346
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>