Stop hiding fully-overloaded inherited rules with custom properties
Rules that both apply directly to an element and are inherited from a
parent are hidden in the styles tab. This is a problem if they contain
custom property declarations, because even if they don't apply to the
element they may be used in value computations in the inherited rules.
This confuses the variable resolution in the styles tab.

Fixed: 41489874
Change-Id: I6429d4651df79bc33190d55b192c14568938a238
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5381087
Auto-Submit: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>