[RPP] Fix not showing entry popover
We have recently introduced a change where we rebuilt the entry popover info only if the entry mouse is hovering over has changed. To check if the entry changed, we store the previous entry index and compare it to the new one.

This introduced a bug where the popover doesn't show up in 2 cases:

1. The 'show hidden entries' arrow was present on an entry and the mouse moved between the arrow and the rest of the entry. In this case the id doesn't change, but the popover should because the arrow popover shows the number of new entries.
2. Navigating to an empty area and then back to the same entry. In this case the popover children were removed, but when we try to show it again, the last entry ID was not reset.

This change resets the ID when popover children are removed and introduces a state that also tracks if the mouse is still hovering over an arrow.

BUG: none
Change-Id: I130c3588d51ca6fdc729529f373cfd78336cee5b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5397418
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>