[RPP] Make EntriesFilter a part of AnnotationsManager
Instead of separately initializing EntriesFilter and AnnotationsManager in TimelinePanel and one depending on another being initialized earlier, make EntriesFilter a part of AnnotationsManager.

The end goal is to make AnnotationsManager responsible for all annotations the user manually makes on a trace and hiding the entries is one of them.

Bug: 313757110
Change-Id: I97df92166839c3e2a54f8335fd1d0f1b0dac1f10
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5389305
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>