Support extension tracks in RPP using user timings
This CLs implements the UI bits to render extension tracks. Namely, gather the data exported by the ExtensionTraceDataHandler and add it to the flame chart.

For this purpose a dedicated track appender is added and called along with the other appenders when the experiment is enabled.

Other changes are made to ensure that UI artifacts in the panel support extension events (i.e. the summary and tree views in the details drawer).

Bug: 40288106
Change-Id: Idd46f58ad5693be48b467b18ea52679acea75573
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5253212
Reviewed-by: Victor Porof <victorporof@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>