Support bigger heap snapshots
I recently received an email about a heap snapshot that couldn't load in
DevTools because it had 227 million edges. The snapshot showed
pathological behavior, but the whole reason the reporter wanted to open
the snapshot was to investigate that pathological behavior, so I think
that DevTools should make a reasonable effort to support such a large
snapshot. We can allow loading that snapshot if we represent
containmentEdges using a new type which might internally use multiple
Uint32Array instances. This makes accesses into the containmentEdges
array a little uglier, but I think that it's overall a worthwhile
change.

When loading a snapshot which is large but not large enough to trigger
the new behavior, I see less than 1% slowdown with this change.

Bug: 332350576
Change-Id: I039ae6b67c43a13b60b35d281f0d04f27bf183c1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5381862
Commit-Queue: Seth Brenith <seth.brenith@microsoft.com>
Reviewed-by: Simon Zünd <szuend@chromium.org>