[LayoutShift] Disable iframe root causes by default
The iframe root cause check casts a very wide net for events that
could have caused an iframe to be added. We are eventually going to drop
the CDP requirement for root causes, so it isn't worth it to make the
process more efficient.

Follow-up to:
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5384002

Lighthouse issue:
https://github.com/GoogleChrome/lighthouse/issues/15869

Bug: 331429737
Change-Id: I68d51ac6b3ed8910ef0b032f885608baf04594b9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5421436
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Adam Raine <asraine@chromium.org>