Allow more nodes in heap snapshots
Shortly after the heap snapshot with too many edges described in
https://crrev.com/c/5381862, I saw another heap snapshot with too many
nodes (100 million, more than the current limit of roughly 76 million).
We can easily apply the same strategy to the nodes array which we are
already using for the containmentEdges array, to allow that even larger
snapshot to load.

Bug: 332350576
Change-Id: Ic5cd3efea08f033ada8518c025f3965e664e1225
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5427134
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Seth Brenith <seth.brenith@microsoft.com>