[sources] Fix VE click events on 'BreakpointMarker' gutter
The 'DebuggerPlugin' currently uses the 'mousedown' event, which is
handled before 'click'. This means that the gutter breakpoint marker
with the VE 'click' handler is removed from the DOM before the 'Click'
can be recorded.

If we use 'click' for toggling gutter breakpoints, the VE works.

R=bmeurer@chromium.org, dsv@chromium.org

Bug: b:330714235
Change-Id: I4688733364236388ac5dfc0d725d66ead108402c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5446904
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>