RPP: add helper to run traceEngine in layout tests
The vast majority of the time we want to test the trace engine in unit
tests. However, we have some axe a11y tests that run in Chromium and
need to be able to construct a trace engine to test the correct things.
Therefore this CL adds a helper to execute the trace engine from a set
of events.

Bug: 40270061
Change-Id: Ib86b9983e4758862ee4b30a17ddb358e3aa9a5ce
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5454232
Commit-Queue: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>