[sources] Don't set identical content twice
For UISourceCodeFrames it can happen that they call 'setDeferredContent'
twice: Once, when the widget itself is shown, triggered in the
'wasShown' callback using the 'lazyContent' passed via constructor.

The second, because the SourcesView manually calls `setUISourceCode`
on the UISourceCodeFrame, even though it already constructed the
same frame with the same UISourceCodeFrame content.

The result is that currently we trigger pretty-printing of large
bundles twice.

This CL adds a sanity check to `setDeferredContent`, to bail if
it was result in an identical `rawContent`.

R=pfaffe@chromium.org

Bug: 333945853
Change-Id: Ie48462421a7fb7f5567530b59b9341b8c72ef582
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5454395
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>