[source-map] Skip parsing if source maps are disabled anyway
The NameResolver currently triggers a full reparse of the generated
script even if JS source maps are disabled. This CL adds an early exit
if source maps are disabled so we don't do an unnecessary parse.

Note that this CL does not change any functionality. We only bail
earlier. Until now, the SourceMapManager would not find a source map
for us to look at, so all the mappings are empty anyway.

R=bmeurer@chromium.org

Bug: 333945853
Change-Id: I20ef25aa7f274267a8000127ec0d47e6982cf66a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5453663
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>