Ensure we dispose of targets when they crash
I found this issue via the attached bug where the performance panel
would freeze. We tracked it down to this line of code:

```
await SDK.TargetManager.TargetManager.instance().resumeAllTargets();
```

Which would never complete, and the UI would never update.

Looking in the protocol monitor, we found that there was a
`Target.targetCrashed` event being emitted. Further investigation
confirmed that if a target crashed, and then you called
`resumeAllTargets`, it would never resume, because the crashed target
would not resume.

To fix this, we listen to the crashed event in the target manager and
dispose of the target when it crashes. This fixes the bug in the
performance panel, because now DevTools can move past the crashed target
and does not attempt to resume it.

Bug: 333989070
Change-Id: I7c128c321ff396ed8bc1c9347ab08e22b003f59c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5458415
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
(cherry picked from commit f61a43042a8433b9487b37813b0bc19494a01dd7)
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5496957
Reviewed-by: Andres Olivares <andoli@chromium.org>