Improve watch expression keyboard navigation.
https://imgur.com/a/lCa8G2B

The keyboard focus on watch expression was not clearly visible and the delete expression button was also not visible causing difficulties navigating and using watch expressions with keyboard. The CL fixes the issue to ensure the keyboard focus is clear and the user is able to navigate to delete button and invoke it.

Bug: 335426677
Change-Id: Ib95ff19356c615622216a5d4d75a79a1b9f2da88
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5464042
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>