Remove TimelineModel.TargetForEvent
Replacing it with a new method that works only on the new engine. To
make the migration easier I added a small wrapper function in
TimelineUIUtils, but once we remove the CompatibleTraceEvent type, we
can remove that too.

Bug: 40287735
Change-Id: I285e8ca8db0c5da4c0e32f7709159d33f92de1c1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5466087
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>