RPP: create fresh recording tracker
We needed to swap out the usage of TimelineModel.isFreshRecording. At
first I thought we could store that status on a trace's metadata, but
then if you loaded a trace that had `isFresh: true` in its saved
JSON...well that just doesn't make any sense :)

So instead, we hold a set of all traces that are fresh in the
FreshRecording.Tracker. This can then be queried with the data (which all
the UI has access to) to find out if the trace is fresh, or not.

Bug: 40287735
Change-Id: Icda8fc4b2bf69d9d7aaf9ccb84ce4400947c5df7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5465575
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>