ARIAUtils.alert should support dialogs
It appears that changes outside of the modal dialog are ignored
by the accessibility implementation in Chromium and therefore alerts triggered when a dialog is opened will not be announced at least with VoiceOver. This PR makes the ARIAUtils.alert detect if a dialog is shown and create an alert within the dialog subtree so that it is considered by a11y tools.

Bug: chromium:24930007
Change-Id: Ic2795a09b9babff99211abe5009b449dc24f7749
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5466282
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>