RPP: extend backendNodeId gathering
This CL adds the ability for the new engine to gather backendNodeIds
from Paint, PaintImage and ScrollLayer events.

It also updates the nodeId type on Paint to be optional, as per the
comment in TimelineModel (which was landed in 2021), making it clear
that we might not always have nodeIds in Paint events.

Bug: 336237398
Change-Id: I0f76c28337fd189695a51cf6e2d2b8eca9a3953e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5472243
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>