RPP: ImagePaintingHandler
In the legacy engine we have the ability to associate DecodeImage events
to PaintImage events. The main reason for this is so if the user clicks a
DecodeImage event, we can show them the right DOM node via the
BackendNodeID, but the BackendNodeID is stored on the PaintImage event,
not the DecodeImage.

Sometimes we can do this directly and go from DecodeImage>PaintImage.

However, other times we need to detour via two other events:
DrawLazyPixelRef and DecodeLaxyPixelRef.

When we see a DrawLazyPixelRef, we associate it to the last PaintImage
event on that thread via the LazyPixelRef key.

When we see a DecodeLaxyPixelRef, we can find its last PaintImage event
by looking up the LazyPixelRef key.

All of this means that when we get a DecodeImage event, we can either
go:

=> DecodeImage > PaintImage

or:

=> DecodeImage > DecodeLaxyPixelRef > PaintImage (via LazyPixelRef key
from the DrawLazyPixelRef event)

Bug: 336237398
Change-Id: I8676d57f30f09719333610b79e9b88635f6e5758
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5471673
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>