RPP: add PageFramesHandler
A very basic handler, but we need to replace the TimelineModel.PageFrame
class, and this is going to do it. We are using this really as a storage
of FrameID=>Frame and nothing more.

Fixed: 336490581
Bug: 40287735
Change-Id: Ied827a34ffd5acc9d4f7c3f493dd9b39ecc1bc5e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5476227
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>